Adding information on the aws_region ENV to make AWS S3 storage work
Coupled with the proposed changes to publify/config/initializers/carrierwave.rb

AWS V4 authentication requires the region to be specifically called or errors will result preventing uploads.  The carrierwave.rb has a proposed change to include an aws_region environment variable; reflecting this proposed change with a proposed change to the README instructions.